pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["arsnovv <arsnov22@mail.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^7.0"
pytest-cov = "^5.0"
 
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'
